[Bench][Pool] Move the baseline-routing test out of the boundary gate
test_benchmark_boundaries.py scans benchmarks/ops for a tests import and a local
gen_inputs. Which baseline a pool case picks is not a boundary, and the file's own
docstring says what it covers, so the test goes to a file named for what it tests.
